pyproject.toml
[tool.poetry]
name = "jmf-llm"
version = "0.1.0"
description = "The JMF-OpenAI-LLM Integration project is a cutting-edge endeavor that brings together the power of OpenAI's large language models (LLM) with the functionality of the JMF Survey App. The project aims to augment the capabilities of the JMF Survey App by leveraging the natural language processing prowess of OpenAI's LLM to providing insights from the survey data."
authors = ["Byamasu Patrick Paul <43136469+byamasu-patrick@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
alembic = "1.12.0"
aniso8601 = "9.0.1"
anyio = "3.7.1"
async-timeout = "4.0.2"
asyncpg = "0.28.0"
atlassian-python-api = "3.39.0"
attrs = "23.1.0"
blinker = "1.6.2"
build = "0.10.0"
CacheControl = "0.13.1"
certifi = "2023.5.7"
cffi = "1.15.1"
chardet = "5.2.0"
charset-normalizer = "3.1.0"
cleo = "2.0.1"
click = "8.1.3"
colorama = "0.4.6"
cryptography = "41.0.1"
distlib = "0.3.7"
exceptiongroup = "1.1.3"
fastapi = "0.103.2"
filelock = "3.12.4"
filetype = "1.2.0"
fire = "0.5.0"
html2text = "2020.1.16"
idna = "3.4"
importlib-metadata = "6.8.0"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpointer = "2.4"
jsonschema = "4.17.3"
keyring = "24.2.0"
langchain = "0.0.310"
langchain-experimental = "0.0.20"
langchainplus-sdk = "0.0.17"
llama-index = "0.8.45.post1"
lxml = "4.9.3"
Mako = "1.2.4"
MarkupSafe = "2.1.3"
marshmallow = "3.19.0"
marshmallow-enum = "1.5.1"
mistune = "3.0.1"
more-itertools = "10.1.0"
msgpack = "1.0.7"
multidict = "6.0.4"
mypy-extensions = "1.0.0"
nest-asyncio = "1.5.8"
nltk = "3.8.1"
numexpr = "2.8.4"
numpy = "1.25.0"
oauthlib = "3.2.2"
openai = "0.27.8"
openapi-schema-pydantic = "1.2.4"
psutil = "5.9.5"
psycopg2 = "2.9.6"
ptyprocess = "0.7.0"
pycparser = "2.21"
pydantic = "1.10.9"
PyJWT = "2.7.0"
pyrsistent = "0.19.3"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-magic = "0.4.27"
pytz = "2023.3"
PyYAML = "6.0"
rapidfuzz = "2.15.2"
referencing = "0.30.2"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
requests-toolbelt = "1.0.0"
rpds-py = "0.10.3"
shellingham = "1.5.3"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.4.1"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.22"}
sse-starlette = "1.6.5"
starlette = "0.27.0"
tabulate = "0.9.0"
tenacity = "8.2.2"
termcolor = "2.3.0"
tiktoken = "0.4.0"
tomli = "2.0.1"
tomlkit = "0.12.1"
trove-classifiers = "2023.9.19"
typing-inspect = "0.8.0"
typing_extensions = "4.5.0"
tzdata = "2023.3"
unstructured = "0.10.15"
urllib3 = "1.26.16"
uvicorn = "0.23.2"
waitress = "2.1.2"
Werkzeug = "2.3.7"
wrapt = "1.15.0"
yarl = "1.9.2"
zipp = "3.17.0"
pgvector = "^0.2.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
